refactor: move `AnalyticsConstants` to :common
Unblocks moving `AnalyticsWidgetProvider` to :widgets

Issue 20737
